Move hasAnyMacro to Lookup.shouldDoEarlyOutput

+++ internal/zinc-core/src/main/scala/sbt/internal/inc/PickleJar.scala
@@ -19,20 +19,18 @@ import java.nio.file.attribute.BasicFileAttributes
 import scala.reflect.io.RootPath
 
 object PickleJar {
-  // create an empty JAR file in case the subproject has no classes.
-  def touch(pickleOut: Path): Unit = {
+  def write(pickleOut: Path, knownProducts: java.util.Set[String]): Unit = {
     if (!Files.exists(pickleOut)) {
       Files.createDirectories(pickleOut.getParent)
       Files.createFile(pickleOut)
     }
-    ()
-  }
-  def write(pickleOut: Path, knownProducts: java.util.Set[String]): Path = {
-    val pj = RootPath(pickleOut, writable = false) // so it doesn't delete the file
-    val result = try Files.walkFileTree(pj.root, deleteUnknowns(knownProducts))
-    finally pj.close()
-    touch(pickleOut)
-    result
+
+    if (!knownProducts.isEmpty) {
+      val pj = RootPath(pickleOut, writable = false) // so it doesn't delete the file
+      try Files.walkFileTree(pj.root, deleteUnknowns(knownProducts))
+      finally pj.close()
+      ()
+    }
   }
 
   def deleteUnknowns(knownProducts: java.util.Set[String]) = new SimpleFileVisitor[Path] {
